tweak rename utils to parse_utils
